refactor(bundle-source): Deduplicate zip-base64 and script bundlers
