+++ packages/bundle-source/test/tool-command.test.js
@@ -3,7 +3,6 @@ import test from '@endo/ses-ava/prepare-endo.js';
 
 import { spawn } from 'child_process';
 import url from 'url';
-import Buffer from 'buffer';
 
 const cwd = url.fileURLToPath(new URL('..', import.meta.url));
